refactor: switch to ts, change from cjs to esm syntax (import instead of require), add typescript dependencies (and openai to be used later)
