+++ package.json
@@ -6,8 +6,12 @@
   "dependencies": {
     "@aws-sdk/client-s3": "^3.100.0",
     "@aws-sdk/lib-storage": "^3.107.0",
+    "@aws-sdk/types": "^3.226.0",
     "@azure/ai-text-analytics": "^5.1.0",
     "@azure/cognitiveservices-computervision": "^8.1.0",
+    "@tsconfig/node16": "^1.0.3",
+    "@types/express": "^4.17.15",
+    "@types/node": "^18.11.18",
     "async": "^3.2.1",
     "axios": "^0.22.0",
     "bcryptjs": "^2.4.3",
@@ -19,12 +23,15 @@
     "express-fileupload": "^1.4.0",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.0.12",
+    "openai": "^3.1.0",
     "react-scripts": "^4.0.3",
-    "synonyms": "^1.0.1"
+    "synonyms": "^1.0.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.4"
   },
   "scripts": {
-    "start": "node app.js",
-    "dev": "nodemon app.js"
+    "start": "node app.ts",
+    "dev": "nodemon app.ts"
   },
   "repository": {
     "type": "git",
