--- routes/userRoute.js
@@ -1,14 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const {
-  registerUser,
-  loginUser,
-  getMe,
-} = require("../controllers/userController.js");
-const { protect } = require("../middleware/authMiddleware");
-
-router.post("/", registerUser);
-router.post("/login", loginUser);
-router.get("/me", protect, getMe);
-
-module.exports = router;
