--- scripts/compileAndSave.js
@@ -1,57 +1,0 @@
-// logic for compiling text and keyphrases from uploaded page and saving in mongoDB document
-
-const { keyPhraseExtraction, getTextFromImage } = require("./textAnalytics");
-
-// extract text and keyPhrase and format as object to be saved in mongoDB
-const textAndPhraseCompile = async (path) => {
-  const { joinedArray, textArray } = await getTextFromImage(path); // returns lines of read text
-  let segmentArray = [];
-
-  for (const line of joinedArray) {
-    let currentPhrase = await keyPhraseExtraction([line]);
-
-    let textSegment = {
-      id: joinedArray.indexOf(line),
-      text: line,
-      phrases: currentPhrase.flat(),
-    };
-
-    segmentArray.push(textSegment);
-  }
-  return { segmentArray, textArray };
-};
-
-const compileAndSave = async (fileName, path, doc) => {
-  let { segmentArray, textArray: rawText } = await textAndPhraseCompile(path);
-  let saveStatus = false;
-  console.log("segmentArray", segmentArray);
-
-  if (
-    segmentArray === undefined ||
-    (segmentArray[0].text === undefined && segmentArray[0].phrases.length === 0)
-  ) {
-    console.log("Error: There is no data to save");
-  } else {
-    let user;
-    try {
-      data = {
-        user: user ? req.user.id : null,
-        page: { fileName, rawText, textByNumber: segmentArray },
-      };
-
-      await doc.create(data);
-      const myDoc = await doc.findOne();
-      if (myDoc) {
-        console.log("Document saved successfully");
-        saveStatus = true;
-      } else {
-        throw new Error("An error occured while attempting to save document");
-      }
-    } catch (err) {
-      console.log(err.stack);
-    }
-    return saveStatus;
-  }
-};
-
-module.exports = { compileAndSave, textAndPhraseCompile };
