+++ tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@tsconfig/node16/tsconfig.json",
+  "compilerOptions": {
+    "noImplicitAny": false,
+    "lib": ["es6"]
+  },
+  "include": [""],
+  "exclude": ["node_modules"]
+}
